remove IMessageBuilder and concrete type
